+++ src/main/kotlin/maryk/core/properties/definitions/TimeDefinition.kt
@@ -9,7 +9,7 @@ import maryk.core.properties.types.TimePrecision
  */
 class TimeDefinition(
         name: String? = null,
-        index: Short = -1,
+        index: Int = -1,
         indexed: Boolean = false,
         searchable: Boolean = true,
         required: Boolean = false,
